moved over the admin file, fixed thomas' stuff with the promotions
